Extract recordableBlogClient
